now I've a function in painter that knows how to render the actor as is

+++ src/main/scala/com/gnosly/fluentsequence/view/fixedwidth/Fixed2DPoint.scala
@@ -1,7 +1,7 @@
 package com.gnosly.fluentsequence.view.fixedwidth
 
 case class Fixed2DPoint(x: Long, y: Long) {
-	def right(i: Int): Fixed2DPoint = Fixed2DPoint(x+i,y)
+	def right(i: Long): Fixed2DPoint = Fixed2DPoint(x+i,y)
 
-	def down(i: Int): Fixed2DPoint = Fixed2DPoint(x,y+i)
+	def down(i: Long): Fixed2DPoint = Fixed2DPoint(x,y+i)
 }
